+++ conf/settings.py
@@ -401,10 +401,10 @@ FEATURE_FLAGS = {
     'INTERNATIONAL_TRIAGE_ON': env.bool('FEATURE_INTERNATIONAL_TRIAGE_ENABLED', False)
 }
 
-# Invest High Potential Opportunities
+# Invest Foreign Direct Investment Opportunities - prev called High Potential Opportunities, hence HPO prefix
 HPO_GOV_NOTIFY_AGENT_EMAIL_ADDRESS = env.str('HPO_GOV_NOTIFY_AGENT_EMAIL_ADDRESS',)
-HPO_GOV_NOTIFY_AGENT_TEMPLATE_ID = env.str('HPO_GOV_NOTIFY_AGENT_TEMPLATE_ID', 'dc467956-f05f-4af9-8713-6c1dc7b7e87a')
-HPO_GOV_NOTIFY_USER_TEMPLATE_ID = env.str('HPO_GOV_NOTIFY_USER_TEMPLATE_ID', '8db76bef-70ec-4d14-97f1-9d087abc4ce2')
+HPO_GOV_NOTIFY_AGENT_TEMPLATE_ID = env.str('HPO_GOV_NOTIFY_AGENT_TEMPLATE_ID', 'c0a46bb9-5f43-4707-9a14-bd5b5675e26e')
+HPO_GOV_NOTIFY_USER_TEMPLATE_ID = env.str('HPO_GOV_NOTIFY_USER_TEMPLATE_ID', 'c7534f85-f975-4a22-904b-5321d7fcb638')
 
 # Directory healthcheck
 DIRECTORY_HEALTHCHECK_TOKEN = env.str('HEALTH_CHECK_TOKEN')
