+++ investment_atlas/tests/test_views.py
@@ -1,7 +1,15 @@
-from unittest.mock import patch
+
+import pytest
+from unittest.mock import call, patch
+from importlib import import_module
+from requests.exceptions import HTTPError
+from django.urls import reverse
+
+from directory_constants import choices
 
 from core.tests.helpers import create_response
-
+from investment_atlas import views
+from investment_atlas.forms import HOW_CAN_WE_HELP_CHOICES, HOW_DID_YOU_HEAR_CHOICES
 from investment_atlas.views import InvestmentOpportunitySearchView
 
 
@@ -782,4 +790,297 @@ def test_planning_status_filter_for_opportunity_search(
     assert view.all_planning_statuses() == [
         ('Planning Status Five', 'Planning Status Five'),
         ('Planning Status Two', 'Planning Status Two'),
+    ]
+
+
+@patch('directory_cms_client.client.cms_api_client.lookup_by_path')
+def test_opportunity_detail(
+    mock_lookup_by_path, settings, client
+):
+    mock_lookup_by_path.return_value = create_response(
+        status_code=200, json_payload={
+            'meta': {'languages': [['en-gb', 'English']]},
+            'page_type': 'InvestmentOpportunityPage',
+        }
+    )
+
+    url = '/international/content/investment/opportunities/rail/'
+
+    response = client.get(url)
+
+    assert response.status_code == 200
+    assert response.template_name == [
+        'investment_atlas/opportunity.html'
+    ]
+
+
+@patch('directory_cms_client.client.cms_api_client.lookup_by_path')
+def test_opportunity_detail_not_found(
+    mock_lookup_by_path, settings, client
+):
+
+    mock_lookup_by_path.return_value = create_response(status_code=404)
+
+    url = '/international/content/investment/opportunities/rail/'
+
+    response = client.get(url)
+
+    assert response.status_code == 404
+
+
+@patch('directory_cms_client.client.cms_api_client.lookup_by_path')
+def test_opportunity_detail_cms_retrieval_ok(
+    mock_lookup_by_path, settings, client
+):
+
+    mock_lookup_by_path.return_value = create_response(
+        status_code=200, json_payload={
+            'title': '1234',
+            'meta': {'languages': [['en-gb', 'English']]},
+            'page_type': 'InvestmentOpportunityPage',
+        }
+    )
+
+    url = '/international/content/investment/opportunities/rail/'
+
+    response = client.get(url)
+
+    assert response.status_code == 200
+    assert response.context_data['page'] == {
+        'title': '1234', 'meta': {'languages': [['en-gb', 'English']]},
+        'page_type': 'InvestmentOpportunityPage',
+    }
+
+
+@patch('directory_cms_client.client.cms_api_client.lookup_by_path')
+def test_opportunity_detail_cms_retrieval_not_ok(
+    mock_lookup_by_path, settings, client
+):
+
+    mock_lookup_by_path.return_value = create_response(status_code=400)
+
+    url = '/international/content/investment/opportunities/rail/'
+
+    with pytest.raises(HTTPError):
+        client.get(url)
+
+
+@patch('directory_cms_client.client.cms_api_client.lookup_by_path')
+def test_foreign_direct_investment_form(mock_lookup_by_path, settings, client):
+    mock_lookup_by_path.return_value = create_response(
+        json_payload={
+            'opportunity_list': [],
+            'meta': {
+                'slug': 'page',
+                'languages': [['en-gb', 'English']],
+            },
+            'page_type': 'ForeignDirectInvestmentFormPage',
+        }
+    )
+
+    url = reverse('fdi-opportunity-request-form')
+
+    response = client.get(url)
+
+    assert response.status_code == 200
+    assert response.template_name == [
+        views.ForeignDirectInvestmentOpportunityFormView.template_name
+    ]
+
+
+@patch('directory_cms_client.client.cms_api_client.lookup_by_path')
+def test_foreign_direct_investment_form_not_found(mock_lookup_by_path, settings, client):
+    mock_lookup_by_path.return_value = create_response(status_code=404)
+
+    url = reverse('fdi-opportunity-request-form')
+    response = client.get(url)
+
+    assert response.status_code == 404
+
+
+@patch('directory_cms_client.client.cms_api_client.lookup_by_path')
+def test_foreign_direct_investment_form_cms_retrieval_ok(mock_lookup_by_path, settings, client):
+    mock_lookup_by_path.return_value = create_response(
+        status_code=200, json_payload={
+            'opportunity_list': [
+                {
+                    # 'pdf_document': 'http://www.example.com/a',  # Disabled for now
+                    'heading': 'some great opportunity',
+                    'meta': {'slug': 'rail', 'url': 'http://www.example.com/test/opp/'}
+                }
+            ],
+            'page_type': 'ForeignDirectInvestmentFormPage',
+        }
+    )
+
+    url = reverse('fdi-opportunity-request-form')
+
+    response = client.get(url)
+
+    assert response.status_code == 200
+    form = response.context_data['form']
+    assert form.fields['opportunities'].choices == [
+        ('http://www.example.com/test/opp/', 'some great opportunity'),
+    ]
+
+
+@patch('directory_cms_client.client.cms_api_client.lookup_by_path')
+def test_foreign_direct_investment_form_cms_retrieval_not_ok(
+    mock_lookup_by_path, settings, client
+):
+    mock_lookup_by_path.return_value = create_response(status_code=400)
+
+    url = reverse('fdi-opportunity-request-form')
+
+    with pytest.raises(HTTPError):
+        client.get(url)
+
+
+@patch('investment_atlas.forms.ForeignDirectInvestmentOpportunityForm.action_class')
+@patch('investment_atlas.forms.ForeignDirectInvestmentOpportunityForm.action_class.save')
+@patch('directory_cms_client.client.cms_api_client.lookup_by_path')
+def test_foreign_direct_investment_form_submmit_cms_retrieval_ok(
+    mock_lookup_by_path, mock_save, mock_action_class, settings, rf,
+    captcha_stub
+):
+    mock_lookup_by_path.return_value = create_response(
+        json_payload={
+            'opportunity_list': [
+                {
+                    # 'pdf_document': 'http://www.example.com/a',  # Disabled for now
+                    'heading': 'some great opportunity',
+                    'meta': {'slug': 'rail', 'url': 'http://www.example.com/test/opp/'}
+                }
+            ],
+            'page_type': 'ForeignDirectInvestmentFormPage',
+        }
+    )
+    mock_save.return_value = create_response(status_code=200)
+    settings.HPO_GOV_NOTIFY_AGENT_EMAIL_ADDRESS = 'invest@example.com'
+
+    url = reverse('fdi-opportunity-request-form')
+
+    request = rf.post(url, data={
+        'given_name': 'Jim',
+        'family_name': 'Example',
+        'job_title': 'Chief chief',
+        'email_address': 'test@example.com',
+        'phone_number': '555',
+        'company_name': 'Example corp',
+        'website_url': 'example.com',
+        'company_address': '123 Some Street, \nSome town, \nSomewhere, \nNarnia',
+        'country': choices.COUNTRY_CHOICES[1][0],
+        'industry': [choices.INDUSTRIES[0][0]],
+        'opportunities': [
+            'http://www.example.com/test/opp/',
+        ],
+        'how_can_we_help': HOW_CAN_WE_HELP_CHOICES[0][0],
+        'your_plans': 'Lorem ipsum dolor sit amet',
+        'how_did_you_hear': HOW_DID_YOU_HEAR_CHOICES[0][0],
+        'email_contact_consent': True,
+        'telephone_contact_consent': True,
+        'g-recaptcha-response': captcha_stub,
+    })
+
+    utm_data = {
+        'campaign_source': 'test_source',
+        'campaign_medium': 'test_medium',
+        'campaign_name': 'test_campaign',
+        'campaign_term': 'test_term',
+        'campaign_content': 'test_content'
+    }
+    request.utm = utm_data
+    request.session = {}
+    response = views.ForeignDirectInvestmentOpportunityFormView.as_view()(
+        request,
+        path='/invest/high-potential-opportunities/contact/success/',
+    )
+
+    assert response.status_code == 302
+    assert response.url == reverse('fdi-opportunity-request-form-success')
+
+    assert mock_action_class.call_args_list[0] == call(
+        email_address=settings.HPO_GOV_NOTIFY_AGENT_EMAIL_ADDRESS,
+        template_id=settings.HPO_GOV_NOTIFY_AGENT_TEMPLATE_ID,
+        form_url=url,
+        sender={'email_address': 'test@example.com', 'country_code': 'AL', 'ip_address': '127.0.0.1'}
+    )
+    assert mock_action_class.call_args_list[1] == call(
+        email_address='test@example.com',
+        template_id=settings.HPO_GOV_NOTIFY_USER_TEMPLATE_ID,
+        form_url=url,
+        email_reply_to_id=settings.HPO_GOV_NOTIFY_USER_REPLY_TO_ID,
+    )
+
+
+@pytest.mark.skip(
+    "Skipped because we're currently not redirecting if there is no session data, "
+    "because we don't need that data until we start passing around PDF download URLs again"
+)
+def test_foreign_direct_investment_form_get_success_page_no_session(client, settings):
+
+    url = reverse('fdi-opportunity-request-form-success')
+
+    response = client.get(url)
+
+    assert response.status_code == 302
+    assert response.url == reverse('fdi-opportunity-request-form')
+
+
+@pytest.mark.skip(
+    "Skipped because we're currently not parking selected opps in session data, "
+    "because we don't need that data until we start passing around PDF download URLs again"
+)
+@patch('directory_cms_client.client.cms_api_client.lookup_by_path')
+def test_foreign_direct_investment_form_get_success_page_with_session(
+    mock_lookup_by_path, settings, client
+):
+
+    settings.SESSION_ENGINE = 'django.contrib.sessions.backends.file'
+    engine = import_module(settings.SESSION_ENGINE)
+    store = engine.SessionStore()
+    store.save()
+    session = store
+    client.cookies[settings.SESSION_COOKIE_NAME] = store.session_key
+
+    session[views.SESSION_KEY_SELECTED_OPPORTUNITIES] = (
+        'http://www.example.com/a'
+    )
+    session.save()
+
+    mock_lookup_by_path.return_value = create_response(
+        json_payload={
+            'opportunity_list': [
+                {
+                    # 'pdf_document': 'http://www.example.com/a',  # Disabled for now
+                    'heading': 'some great opportunity',
+                    'meta': {'slug': 'rail', 'url': 'http://www.example.com/test/opp/'}
+                },
+                {
+                    # 'pdf_document': 'http://www.example.com/b',  # Disabled for now
+                    'heading': 'some other opportunity',
+                    'meta': {'slug': 'other', 'url': 'http://www.example.com/test/other-opp/'}
+                }
+            ],
+            'meta': {
+                'slug': 'page',
+                'languages': [['en-gb', 'English']],
+            },
+            'page_type': 'ForeignDirectInvestmentFormSuccessPage',
+        }
+    )
+
+    url = reverse('fdi-opportunity-request-form-success')
+
+    response = client.get(url)
+
+    assert response.status_code == 200
+    assert response.context_data['page']
+
+    assert response.context_data['opportunities'] == [
+        {
+            'pdf_document': 'http://www.example.com/a',
+            'heading': 'some great opportunity',
+            'meta': {'slug': 'rail', 'url': 'http://www.example.com/test/opp/'}
+        }
     ]
